[REF] shopfloor_reception_packaging_dimension_mobile: extract packaging dimension form edit in a component
Extract the form to edit packaging dimensions in a component so that it can be reused
